schema edits

+++ Controllers/orderController.js
@@ -3,6 +3,9 @@ const Order = require("../Models/orderModel")
 // Get all cart
 exports.getAllOrders = async (req, res) => {
     try {
+        if (req.user.role === 'user') {
+            return res.status(401).json({ message: "not autorized" })
+        }
         const order = await Order.find();
         if (!order) {
             return res.status(404).json({ message: "order not found" });
@@ -13,51 +16,28 @@ exports.getAllOrders = async (req, res) => {
         console.log(err);
     }
 }
+
+exports.getOrderHistoryByUserId = async (req, res) => {
+    try {
+        const order = await Order.find({ user_id: req.user._id });
+        if (!order) {
+            return res.status(404).json({ message: "order not found" });
+        }
+        res.status(200).send({ message: true, data: order });
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+        console.log(err);
+    }
+}
+
 // add order
 exports.addOrder = async (req, res) => {
     try {
-        const {
-            user_id,
-            cart_id,
-            shippingAddress: {
-                address,
-                city,
-                postalCode,
-                country
-            },
-            paymentMethod,
-            // paymentResult: {
-            //     id,
-            //     status, update_time,
-            //     email_address
-            // },
-            totalPrice,
-            isPaid,
-            paidAt,
-            isDelivered,
-            deliverdAt
-        } = req.body;
-
+        if (req.user.role !== "user") {
+            return res.status(401).json({ message: "not authorized" })
+        }
         const addNewOrder = await Order.create({
-            user_id,
-            cart_id,
-            shippingAddress: {
-                address,
-                city,
-                postalCode,
-                country
-            },
-            paymentMethod,
-            // paymentResult: {
-            //     id,
-            //     status, update_time,
-            //     email_address
-            // },
-            totalPrice,
-            isPaid,
-            paidAt,
-            isDelivered,
-            deliverdAt
+            user_id: req.user._id, ...req.body
         });
 
         if (!addNewOrder) {
@@ -71,10 +51,14 @@ exports.addOrder = async (req, res) => {
         console.log(err);
     }
 };
+
 // Edit cart
 exports.editOneOrder = async (req, res) => {
     try {
         let { id } = req.params;
+        if (req.user.role == user) {
+            return res.status(401).json({ message: 'not authorized' })
+        }
         let body = req.body;
         const updateOrder = await Order.updateOne(
             { _id: id },
@@ -95,6 +79,9 @@ exports.editOneOrder = async (req, res) => {
 //Delete One Cart
 exports.deleteOrder = async (req, res) => {
     try {
+        if (req.user.role == user) {
+            return res.status(401).json({ message: 'not authorized' })
+        }
         let { id } = req.params;
         const deleteOneOrder = await Order.findByIdAndDelete(
             { _id: id }
